[WIP] Refactor common functions in public/private data interface
